// AllTheBeans-Frontend/src/types/models/coffeeBean.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * AllTheBeans.API
 * OpenAPI spec version: 1.0
 */

export interface CoffeeBean {
  id?: number;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  image?: string | null;
  /** @nullable */
  colour?: string | null;
  /** @nullable */
  country?: string | null;
  cost?: number;
}